Change Games to be created from a json file
This is a huge change. The goal of this change is to be able to add
games and fix simple things without updating the mod.
Changes:
- Removed all classes extending Game.
- Added games.json which holds the information contained in the removed
classes.
- Added UpToDateThread to extract the games.json file from jar and to
read it. (This will be able to download the latest version in the
future.)
- Added GameConfiguration which represents the "games" elements of the
games.json.
- Added GameMangaer which holds all GameConfigurations contained in
games.json. It is able to create Game instances.
- Added ComponentsManager which is used to instantiate the components of
a Game.
- Added ComponentInstanceConfiguration which represents the "components"
part of games.json and is needed to instantiate Components with a
non-default constructor.
- Added a replacement GameType class which only contains the id the mod
uses for games. The GameType of the Public API will be removed.
- Changed Game to match to the new system. Removed all fields which are
now in games.json and added a GameConfiguration field. Also added the
constant NO_GAME which will be used instead of null to indicate that
there is no game.
- Adjusted the Detectors to work with the new System and NO_GAME.
- Fixed other problems which occurred with changes.

Currently Broken / Missing:
- Painball cooldown dependant killstreaks and saving the meassured time
between matches.
- Configuration integration. (All components are currently enabled)
- HudPixel UpToDate to download new versions of the games.json